switched to double precision

+++ Model3DViewer/FractalViewer.h
@@ -13,7 +13,7 @@ public:
 private:
 	static void loadShader();
 	static GLuint shaderID, ratioPos, zoomPos, offXPos, offYPos;
-	static float zoom, offX, offY;
+	static double zoom, offX, offY;
 	static int _windowID, _width, _height;
 	static void display();
 	static void reshape(int width, int height);
